Restructured projects feature into smaller parts for cleaner code purposes
