+++ src/components/loading-states/error-state-component.tsx
@@ -1,9 +1,6 @@
 function ErrorStateComponent({ errorMessage }: { errorMessage: string }) {
   return (
-    <p className="loading-status text-red text-align-center">
-      {" "}
-      {errorMessage}{" "}
-    </p>
+    <p className="loading-status text-red text-align-center">{errorMessage}</p>
   );
 }
