+++ src/features/projects/providers/projects-table-provider.tsx
@@ -2,7 +2,7 @@
 
 import { Provider } from "react-redux";
 import { store } from "@/redux/store";
-import ProjectsTable from "../../../components/table/tables/project-table/projects-table";
+import ProjectsTable from "../components/project-table/projects-table";
 
 function ProjectsTableProvider() {
   return (
